inclusions in sweetalert and toastr

+++ app/Http/Controllers/CategoryCotroller.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use RealRashid\SweetAlert\Facades\Alert;
 
 class CategoryCotroller extends Controller
 {
@@ -14,9 +16,12 @@ class CategoryCotroller extends Controller
      */
     public function index()
     {
-        $categories = category::all();
+        $categories = category::with('parent')->get();
+        if (session('success_message')) {
+            Alert::toast(session('success_message'), 'success')->autoClose(4000);
+        }
 
-        return view('category.index',compact('categories'));
+        return view('category.index', compact('categories'));
     }
 
     /**
@@ -24,31 +29,34 @@ class CategoryCotroller extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
+    public function create(category $category)
     {
-        return view('category.create');
+        $mains = DB::table('categories')
+            ->whereNull('parent_category_id')
+            ->get();
+        return view('category.create', compact('mains'));
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
         category::create([
-            'name'=> $request->input('name'),
-            'description'=> $request->input('description'),
-            'parent_category_id'=>$request->input('parent_category_id')
+            'name' => $request->input('name'),
+            'description' => $request->input('description'),
+            'parent_category_id' => $request->input('parent_category_id')
         ]);
-        return redirect()->route('category.index');
+        return redirect()->route('category.index')->withSuccessMessage('category created successfully');
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\category  $category
+     * @param \App\Models\category $category
      * @return \Illuminate\Http\Response
      */
     public function show(category $category)
@@ -59,41 +67,49 @@ class CategoryCotroller extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\category  $category
+     * @param \App\Models\category $category
      * @return \Illuminate\Http\Response
      */
     public function edit(category $category)
     {
-        return view('category.edit',compact('category'));
+        $mains = category::with('child')->whereNull('parent_category_id')->get();
+        return view('category.edit', compact('category', 'mains'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\category  $category
+     * @param \Illuminate\Http\Request $request
+     * @param \App\Models\category $category
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, category $category)
     {
+        if ($request->parent_category_id==$category->id){
+            return redirect()->back()->withToastError('child cannot be parent');
+        }
         $category->update([
-            'name'=> $request->input('name'),
-            'description'=> $request->input('description'),
-            'parent_category_id'=>$request->input('parent_category_id')
+            'name' => $request->input('name'),
+            'description' => $request->input('description'),
+            'parent_category_id' => $request->input('parent_category_id')
         ]);
-        return redirect()->route('category.index');
+        return redirect()->route('category.index')->withSuccessMessage('category updated successfully');
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\category  $category
+     * @param \App\Models\category $category
      * @return \Illuminate\Http\Response
      */
     public function destroy(category $category)
     {
-        $category->delete();
+        if ($category->child()->exists()) {
+            return redirect()->route('category.index')->withToastWarning('cannot delete existing parent category');
+        } else {
+            $category->delete();
+            return redirect()->route('category.index')->withSuccessMessage('category deleted successfully');
+        }
 
-        return redirect()->route('category.index');
     }
 }
